third_party/blink/web_tests/external/wpt/WebCryptoAPI/generateKey/failures_X448.https.any.worker-expected.txt

This is a testharness.js-based test.
Found 32 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Bad usages: generateKey({name: X448}, true, [encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, deriveKey, encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, deriveBits, deriveKey, deriveBits, deriveKey, deriveBits, encrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, deriveKey, decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, deriveBits, deriveKey, deriveBits, deriveKey, deriveBits, decrypt])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [sign])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, sign])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, deriveKey, sign])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, sign])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, deriveBits, deriveKey, deriveBits, deriveKey, deriveBits, sign])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [verify])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, verify])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, deriveKey, verify])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, verify])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, deriveBits, deriveKey, deriveBits, deriveKey, deriveBits, verify])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, deriveKey, wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, deriveBits, deriveKey, deriveBits, deriveKey, deriveBits, wrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, deriveKey, unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveBits, unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Bad usages: generateKey({name: X448}, true, [deriveKey, deriveBits, deriveKey, deriveBits, deriveKey, deriveBits, unwrapKey])
  assert_equals: Bad usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Empty usages: generateKey({name: X448}, false, [])
  assert_equals: Empty usages not supported expected "SyntaxError" but got "NotSupportedError"
[FAIL] Empty usages: generateKey({name: X448}, true, [])
  assert_equals: Empty usages not supported expected "SyntaxError" but got "NotSupportedError"
Harness: the test ran to completion.
